requirements.txt
asgiref==3.8.1
awscli==1.33.4
backports.zoneinfo==0.2.1
boto3==1.34.103
botocore==1.34.122
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
django-filter==24.2
django-s3-sqlite==0.0.5
django-storages==1.14.4
djangorestframework==3.15.1
docutils==0.16
jmespath==1.0.1
pillow==10.4.0
pyasn1==0.6.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==1.26.20
